feat: Add squirel, add Mcq, fix the shaking bug during flat running
- Introduce a squirrel character, along with nuts and a vine.

- Implement an MCQ feature for the enigma, enabling the squirrel to
pose questions to the player. This interaction allows the squirrel to
ask questions, player to collect nuts, and if all nuts are gathered,
a vine will deploy, enabling the player to climb a cliff.

- Address an issue where the player experiences shaking while running
on flat ground. The problem appears to stem from the normal nudge
factor occasionally reducing movement to zero. The solution was to
introduce a slight, imperceptible slope to prevent this issue.